requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
motor==3.7.1
pymongo==4.15.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
playwright==1.55.0
openai==1.102.0
pytest==8.4.2
